+++ app/src/main/java/crux/bphc/cms/models/Download.kt
@@ -1,0 +1,9 @@
+package crux.bphc.cms.models
+
+import java.io.File
+
+data class Download(
+        val file: File,
+        val description: String,
+        val iconResource: Int
+)